+++ WorkingHoursMicroservice/src/main/java/com/medimate/WorkingHoursMicroservice/services/DoctorService.java
@@ -1,5 +1,6 @@
 package com.medimate.WorkingHoursMicroservice.services;
 
+import com.medimate.WorkingHoursMicroservice.models.Admin;
 import com.medimate.WorkingHoursMicroservice.models.Doctor;
 import com.medimate.WorkingHoursMicroservice.repositories.DoctorRepository;
 import com.medimate.WorkingHoursMicroservice.viewmodels.DoctorVM;
@@ -12,27 +13,32 @@ import java.util.Optional;
 @Service
 public class DoctorService {
 
-  @Autowired
-  DoctorRepository repo;
+    @Autowired
+    DoctorRepository repo;
 
 
-  public Doctor addOne(DoctorVM doctorVM){
-      return repo.save(
-              new Doctor(
-                      doctorVM.getFirstName(),
-                      doctorVM.getLastName(),
-                      doctorVM.getTitle()
-              )
-      );
-  }
+    public Doctor addOne(DoctorVM doctorVM) {
+        return repo.save(
+                new Doctor(
+                        doctorVM.getFirstName(),
+                        doctorVM.getLastName(),
+                        doctorVM.getTitle()
+                )
+        );
+    }
 
-  public void deleteById(Integer id){
-      repo.deleteById(id);
-  }
+    public void deleteById(Integer id) {
+        Optional<Doctor> doctorOptional = repo.findById(id);
+        if (doctorOptional.isPresent()) {
+            repo.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("Could not find doctor with id " + id);
+        }
+    }
 
-  public Doctor getById(Integer id){
-      Optional<Doctor> doctorOptional = repo.findById(id);
-      return doctorOptional.orElseThrow(() -> new EntityNotFoundException("Could not find doctor with id %d".formatted(id)));
-  }
+    public Doctor getById(Integer id) {
+        Optional<Doctor> doctorOptional = repo.findById(id);
+        return doctorOptional.orElseThrow(() -> new EntityNotFoundException("Could not find doctor with id %d".formatted(id)));
+    }
 
 }
